hotfix: 기본정보 require 추가 및 휴무여부 초기화 설정
